rearange image-processing package Moved cut houses to image_processing. Changed parametrized folder to core and moved all images to one folder
