--- cutHouseFaces/Luftbilder/Info.txt
@@ -1,2 +1,0 @@
-UTM32 Dateiname: 32EEE_NNNN
-Referenzpunkt links unten
