+++ image-processing/core/detection.py
