+++ image-processing/cut_houses/__init__.py
