+++ image-processing/cut_houses/cut_houses.py
