+++ image-processing/img_results/02_histogramm_spreading.png
