+++ image-processing/img_results/03_quantisierung_1_bit.png
