+++ image-processing/img_results/07_no_shadows_greens.png
