+++ image-processing/img_results/08_no_shadows_histogramm_spreading.png
