+++ image-processing/img_results/09_no_shadows_quanisirung_1_bit.png
