+++ image-processing/img_results/comparisons/13_no_shadows_greens_by_hand.png
