+++ image-processing/img_results/comparisons/2_only_greens.png
