+++ image-processing/img_results/comparisons/3_no_shadows_greens.png
