+++ image-processing/img_results/testing/00_input_corrected.png
