+++ image-processing/img_results/testing/32692_5347_corrected.png
