+++ image-processing/main.py
@@ -2,7 +2,7 @@ import v1.green_detector
 import v2.shadow_corrected_green_detector
 import v3.noise_reduction_green_detector
 import optimize
-import parameterized.configurator as config
+import core.configurator as config
 
 folder = "testing/"
 data = [
@@ -16,21 +16,19 @@ numer_of_results: int = 5
 
 
 if __name__ == '__main__':
-    #v1.green_detector.process_image("00_input.tif")
-    #v2.shadow_corrected_green_detector.process_image("00_input.tif")
-    #v3.noise_reduction_green_detector.process_image("00_input.tif")
+    # v1.green_detector.process_image("00_input.tif")
+    # v2.shadow_corrected_green_detector.process_image("00_input.tif")
+    # v3.noise_reduction_green_detector.process_image("00_input.tif")
 
-    #default = config.get_default_configuration()
-    #defaultRunner = config.create_process(default)
-    #res = defaultRunner(f'{folder}00_input.png', None)
-
+    # default = config.get_default_configuration()
+    # defaultRunner = config.create_process(default)
+    # res = defaultRunner(f'{folder}00_input.png', None)
 
     best = optimize.run(data,
-                 population_size=population_size,
-                 mutation_chance=mutation_chance,
-                 generations=generations,
-                 numer_of_results=numer_of_results)
+                        population_size=population_size,
+                        mutation_chance=mutation_chance,
+                        generations=generations,
+                        numer_of_results=numer_of_results)
 
     for i, config in enumerate(best):
         print(f'Position {i+1}:\n{config}\n')
-
